Split up calendar code into separate files.
Now `calendar.js` is "only" 282 lines long! cc @noahmanger
