+++ fec/fec/static/js/calendar.js
@@ -4,125 +4,25 @@ var $ = require('jquery');
 var URI = require('urijs');
 var _ = require('underscore');
 var moment = require('moment');
+require('fullcalendar');
 
 var urls = require('fec-style/js/urls');
 var dropdown = require('fec-style/js/dropdowns');
 var Handlebars = require('hbsfy/runtime');
 var helpers = require('fec-style/js/helpers');
 
-Handlebars.registerHelper(helpers.helpers);
+var calendarTooltip = require('./calendar-tooltip');
+var calendarHelpers = require('./calendar-helpers');
+require('./calendar-list-view');
 
-require('fullcalendar');
+Handlebars.registerHelper(helpers.helpers);
 
 var templates = {
   details: require('../hbs/calendar/details.hbs'),
   download: require('../hbs/calendar/download.hbs'),
   subscribe: require('../hbs/calendar/subscribe.hbs'),
-  events: require('../hbs/calendar/events.hbs'),
   listToggles: require('../hbs/calendar/listToggles.hbs')
 };
-
-function Listeners() {
-  this.listeners = [];
-}
-
-Listeners.prototype.on = function($elm) {
-  var args = _.toArray(arguments).slice(1);
-  this.listeners = this._listeners || [];
-  this.listeners.push({$elm: $elm, args: args});
-  $elm.on.apply($elm, args);
-};
-
-Listeners.prototype.off = function() {
-  this.listeners.forEach(function(listener) {
-    var $elm = listener.$elm;
-    var args = listener.args;
-    $elm.off.apply($elm, args);
-  });
-};
-
-var FC = $.fullCalendar;
-var View = FC.View;
-
-var categories = {
-  Elections: ['election'],
-  Deadlines: ['report', 'ie', 'ec'],
-  Outreach: ['roundtables', 'conferences'],
-  Meetings: ['open', 'executive'],
-  Rules: ['aos'],
-  Other: ['litigation', 'fea']
-};
-
-var categoriesInverse = _.reduce(_.pairs(categories), function(memo, pair) {
-  var key = pair[0];
-  var values = pair[1];
-  _.each(values, function(value) {
-    memo[value] = key;
-  });
-  return memo;
-}, {});
-
-var categoryGroups = function(events, start, end) {
-  return _.chain(events)
-    .filter(function(event) {
-      return start <= event.start && event.start < end;
-    })
-    .sortBy('start')
-    .groupBy(function(event) {
-      var category = event.category ? event.category.split(/[ -]+/)[0].toLowerCase() : null;
-      return categoriesInverse[category];
-    })
-    .map(function(values, key) {
-      return {
-        title: key,
-        events: values
-      };
-    })
-    .sortBy(function(group) {
-      return Object.keys(categories).indexOf(group.title);
-    })
-    .value();
-};
-
-var chronologicalGroups = function(events, start, end) {
-  events = _.chain(events)
-    .filter(function(event) {
-      return start <= event.start && event.start < end;
-    })
-    .sortBy('start')
-    .value();
-  return [{events: events}];
-};
-
-var ListView = View.extend({
-  setDate: function(date) {
-    var intervalUnit = this.options.duration.intervalUnit || this.intervalUnit;
-    View.prototype.setDate.call(this, date.startOf(intervalUnit));
-  },
-
-  renderEvents: function(events) {
-    var groups = this.options.categories ?
-      categoryGroups(events, this.start, this.end) :
-      chronologicalGroups(events, this.start, this.end);
-    var settings = {
-      duration: this.options.duration.intervalUnit,
-      sortBy: this.options.sortBy
-    };
-    this.el.html(templates.events({groups: groups, settings: settings}));
-    this.dropdowns = $(this.el.html).find('.dropdown').map(function(idx, elm) {
-      return new dropdown.Dropdown($(elm), {checkboxes: false});
-    });
-  },
-
-  unrenderEvents: function() {
-    this.dropdowns.each(function(idx, dropdown) {
-      dropdown.destroy();
-    });
-    this.el.html('');
-  }
-});
-
-FC.views.list = ListView;
 
 var LIST_VIEWS = ['quarterTime', 'quarterCategory', 'monthTime', 'monthCategory'];
 
@@ -255,11 +155,11 @@ Calendar.prototype.success = function(response) {
       start: event.start_date ? moment.utc(event.start_date) : null,
       end: event.end_date ? moment.utc(event.end_date) : null,
       allDay: moment.utc(event.start_date).format('HHmmss') === '000000' && event.end_date === null,
-      className: getEventClass(event),
+      className: calendarHelpers.getEventClass(event),
       detailUrl: event.url
     };
     _.extend(processed, {
-      google: getGoogleUrl(processed),
+      google: calendarHelpers.getGoogleUrl(processed),
       download: self.exportUrl.clone().addQuery({event_id: event.event_id}).toString()
     });
     return processed;
@@ -342,7 +242,7 @@ Calendar.prototype.handleEventClick = function(calEvent, jsEvent, view) {
   var $target = $(jsEvent.target);
   if (!$target.closest('.tooltip').length) {
     var $eventContainer = $target.closest('.fc-content');
-    var tooltip = new CalendarTooltip(
+    var tooltip = new calendarTooltip.CalendarTooltip(
       templates.details(_.extend({}, calEvent, {detailsId: this.detailsId})),
       $eventContainer.parent()
     );
@@ -379,85 +279,4 @@ Calendar.prototype.highlightToday = function() {
     .addClass('fc-today');
 };
 
-var classMap = {
-  aos: 'fc--rules',
-  election: 'fc--election',
-  report: 'fc--deadline',
-  open: 'fc--meeting',
-  executive: 'fc--meeting',
-  roundtables: 'fc--outreach',
-  conferences: 'fc--outreach',
-  litigation: 'fc--other',
-  fea: 'fc--other',
-  ie: 'fc--deadline',
-  ec: 'fc--deadline'
-};
-
-function getEventClass(event) {
-  var className = '';
-  var category = event.category ? event.category.split(/[ -]+/)[0] : null;
-  className += moment.utc(event.start_date).format('HHmmss') === '000000' ? 'fc--allday' : '';
-  className += category ? ' ' + classMap[category.toLowerCase()] : '';
-  return className;
-}
-
-function getGoogleUrl(event) {
-  var fmt, dates;
-  if (event.end) {
-    fmt = 'YYYYMMDD[T]HHmmss';
-    dates = event.start.format(fmt) + '/' + event.end.format(fmt);
-  } else {
-    fmt = 'YYYYMMDD';
-    dates = event.start.format(fmt) + '/' + event.start.clone().add(1, 'day').format(fmt);
-  }
-  return URI('https://calendar.google.com/calendar/render')
-    .addQuery({
-      action: 'TEMPLATE',
-      text: event.title,
-      details: event.summary,
-      dates: dates
-    })
-    .toString();
-}
-
-function getUrl(path, params) {
-  return URI(window.API_LOCATION)
-    .path(Array.prototype.concat(window.API_VERSION, path || [], '').join('/'))
-    .addQuery({
-      api_key: window.API_KEY,
-      per_page: 500
-    })
-    .addQuery(params || {})
-    .toString();
-}
-
-function CalendarTooltip(content, $container) {
-  this.$content = $(content);
-  this.$container = $container;
-  this.$close = this.$content.find('.js-close');
-  this.$dropdown = this.$content.find('.dropdown');
-  this.exportDropdown = new dropdown.Dropdown(this.$dropdown, {checkboxes: false});
-
-  this.events = new Listeners();
-  this.events.on(this.$close, 'click', this.close.bind(this));
-  this.events.on($(document.body), 'click', this.handleClickAway.bind(this));
-}
-
-CalendarTooltip.prototype.handleClickAway = function(e) {
-  var $target = $(e.target);
-  if (!this.$content.has($target).length && !this.$container.has($target).length) {
-    this.close();
-  }
-};
-
-CalendarTooltip.prototype.close = function() {
-  this.$content.remove();
-  this.exportDropdown.destroy();
-  this.$container.find('.fc-content').focus();
-  this.events.off();
-};
-
-module.exports = {
-  Calendar: Calendar,
-  getUrl: getUrl
-};
+module.exports = {Calendar: Calendar};
